Use integer arguments directly in Pageable.ofSize() and withPage() methods in AuditServiceImpl

+++ src/main/java/org/saidone/service/audit/AuditServiceImpl.java
@@ -79,7 +79,7 @@ public class AuditServiceImpl extends BaseComponent implements AuditService {
         if (!criteriaList.isEmpty()) {
             query.addCriteria(new Criteria().andOperator(criteriaList.toArray(new Criteria[0])));
         }
-        var pageable = Pageable.ofSize((int) maxItems).withPage((int) skipCount);
+        var pageable = Pageable.ofSize(maxItems).withPage(skipCount);
         query.with(pageable);
         query.with(Sort.by(Sort.Direction.DESC, "timestamp"));
         return mongoTemplate.find(query, AuditEntry.class);
